move EditorViewState to src/vs/workbench/browser/quickaccess.ts
